Keep chat stream handler tests easier to extend
The chat stream handler tests carried their SSE collector and mock stream builders inline, so this extracts the repeated test scaffolding into a sibling helper module without changing stream assertions or coverage.

Constraint: Refactor must preserve chat stream handler test behavior and SSE expectations exactly
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Collapse test cases into parameterized fixtures | would obscure the stream-sequence intent that each case documents
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep shared chat stream test scaffolding in the helper module unless a case needs truly one-off inline setup
Tested: deno test --no-check --allow-all src/agent/runtime/chat-stream-handler.test.ts; deno fmt --check src/agent/runtime/chat-stream-handler.test.ts src/agent/runtime/chat-stream-handler.test-helpers.ts; deno lint src/agent/runtime/chat-stream-handler.test.ts src/agent/runtime/chat-stream-handler.test-helpers.ts; git diff --check
Not-tested: Full veryfront-code repo suite
